+++ lib/src/provider/theme_provider.dart
@@ -5,99 +5,49 @@ import '../controller/theme_controller.dart';
 import '../controller/theme_command.dart';
 import 'inherited_theme.dart';
 
-/// Signature for a function that returns the current theme.
-///
-/// Used by [ThemeProvider].
-typedef Widget ThemedAppBuilder(ThemeData themeData);
-
 /// Wrap [MaterialApp] in [ThemeProvider] to get theme functionalities.
+/// You may wrap seperate parts of the app with multiple [ThemeProvider]s
+/// to use multiple theme sections across the app.
 class ThemeProvider extends StatelessWidget {
-  /// Called to obtain the child app.
-  ///
-  /// [themeData] provided refers to the current theme.
-  /// Use it to provide theme data to the [MaterialApp] as,
-  /// ```dart
-  /// ThemeProvider(
-  ///   builder: (_, theme) => MaterialApp(
-  ///         theme: theme,
-  ///         home: HomePage(),
-  ///       ),
-  /// )
-  /// ```
-  ///
-  /// This will ensure that app gets rebuilt when the theme changes.
-  final ThemedAppBuilder builder;
+  /// The widget below this widget in the tree.
+  final Widget child;
 
-  /// [defaultThemeId] is optional.
-  /// If not provided, default theme set by [ThemeController] to be the first provided theme.
-  /// Otherwise the given theme will be set as the default theme.
-  /// [AssertionError] will be thrown if there is no theme with [defaultThemeId].
+  /// Optional field which will set the default theme out of themes provided in [themes].
+  /// If not provided, default theme will be the first provided theme.
   final String defaultThemeId;
 
-  /// List of [AppTheme]s. This list will be passed to the
-  /// [ThemeController] when widget is built.
+  /// List of themes to be available for child listners.
+  /// If [themes] are not supplies [AppTheme.light()] and [AppTheme.dark()] is assumed.
+  /// If [themes] are supplied, there have to be at least 2 [AppTheme] objects inside the list.
   final List<AppTheme> themes;
 
-  /// [saveThemesOnChange] refers to whether to persist the theme on change.
-  /// If it is `true`, theme will be saved to disk whenever the theme changes.
+  /// Whether to persist the theme on change.
+  /// If `true`, theme will be saved to disk whenever the theme changes.
+  /// By default this is `false`.
   final bool saveThemesOnChange;
 
   /// Whether to load the theme on initialization.
-  /// If this is true, default onInitCallback will be executed instead.
+  /// If `true`, default [onInitCallback] will be executed instead.
   final bool loadThemeOnInit;
 
-  /// [onInitCallback] is the callback which is called when the ThemeController is first initialed.
-  /// You can use this to call `controller.loadThemeById(ID)` or equalent to set theme.
+  /// The callback which is to be called when the [ThemeController] is first initialed.
   final ThemeControllerHandler onInitCallback;
 
   /// Theme provider id to distinguish between ThemeProviders.
+  /// Provide distinct values if you intend to use multiple theme providers.
   final String providerId;
 
   /// Creates a [ThemeProvider].
   /// Wrap [MaterialApp] in [ThemeProvider] to get theme functionalities.
-  /// Usage example:
-  /// ```dart
-  /// ThemeProvider(
-  ///   builder: (_, theme) => MaterialApp(
-  ///         theme: theme,
-  ///         home: HomePage(),
-  ///       ),
-  /// )
-  /// ```
-  ///
-  /// If [themes] are not supplies [AppTheme.light()] and [AppTheme.dark()] is assumed.
-  ///
-  /// If [themes] are supplied, there have to be at least 2 [AppTheme] objects inside
-  /// the list. Otherwise an [AssertionError] is thrown.
-  ///
-  /// [defaultThemeId] can also be provided to override the default theme.
-  /// (which is the first theme in the [themes] list)
-  /// **[AppTheme]s must not have conflicting theme ids.**
-  ///
-  /// [saveThemesOnChange] refers to whether to persist the theme on change.
-  /// On default this is `false`.
-  /// If it is `true`, theme will be saved to disk whenever the theme changes.
-  /// **If you use this, do NOT use nested [ThemeProvider]s as all will be saved in the same key**
-  ///
-  /// [loadThemesOnStartup] refers to whether to load the theme when the controller is initialized.
-  /// If `true`, this will load the default theme provided (or the first theme if default is `null`)
-  /// and then asynchronously load the persisted theme.
-  /// If no persisted theme found, the theme will remain as the default one.
-  /// By default this is `false`
-  ///
-  /// [loadThemeOnInit] will load a previously saved theme from disk.
-  /// If [loadThemeOnInit] is provided, [onInitCallback] will be ignored.
-  /// So [onInitCallback] and [loadThemeOnInit] can't both be provided at the same time.
-  ///
-  /// Theme provider id to distinguish between ThemeProviders.
-  /// Use this to add multiple themeproviders to your app.
+  /// You may wrap seperate parts of the app with multiple [ThemeProvider]s
+  /// to use multiple theme sections across the app.
   ThemeProvider({
     Key key,
     this.providerId = "default",
     List<AppTheme> themes,
     this.defaultThemeId,
     this.onInitCallback,
-    @required this.builder,
+    @required this.child,
     this.saveThemesOnChange = false,
     this.loadThemeOnInit = false,
   })  : this.themes = themes ?? [AppTheme.light(), AppTheme.dark()],
@@ -105,25 +55,21 @@ class ThemeProvider extends StatelessWidget {
     assert(this.themes.length >= 2, "Theme list must have at least 2 themes.");
   }
 
-  /// Obtains the nearest [ThemeController] up its widget tree and
-  /// returns its value as a [ThemeCommand].
-  ///
-  /// Gets the reference to [ThemeController] but as a reduced version.
+  /// Gives reference to a [ThemeCommand] of the nearest [ThemeProvider] up the widget tree
+  /// and will provide commands to change the theme.
   static ThemeCommand controllerOf(BuildContext context) {
     return InheritedThemeController.of(context);
   }
 
-  /// Returns the options passed by the [ThemeProvider].
+  /// Returns the options passed by the nearest [ThemeProvider] up the widget tree.
   /// Call as `ThemeProvider.optionsOf<ColorClass>(context)` to get the
   /// returned object casted to the required type.
   static T optionsOf<T>(BuildContext context) {
     return InheritedThemeController.of(context).theme.options as T;
   }
 
-  /// Returns the current app theme passed by the [ThemeProvider].
-  ///
-  /// Call as `ThemeProvider.themeOf(context).data` to get [ThemeData].
-  /// This is same as `Theme.of(context)` under a theme provider.
+  /// Returns the current app theme passed by the nearest [ThemeProvider] up the widget tree.
+  /// Use as `ThemeProvider.themeOf(context).data` to get [ThemeData].
   static AppTheme themeOf(BuildContext context) {
     return InheritedThemeController.of(context).theme;
   }
@@ -139,9 +85,7 @@ class ThemeProvider extends StatelessWidget {
         loadThemeOnInit: loadThemeOnInit,
         saveThemesOnChange: saveThemesOnChange,
       ),
-      child: Builder(
-        builder: (context) => builder(ThemeProvider.themeOf(context).data),
-      ),
+      child: child,
     );
   }
 }